Refactoring of view classes to view folder
