+++ TrainingApp/app/src/main/java/com/traininapp/View/StatisticCard.java
@@ -1,4 +1,6 @@
-package com.traininapp.viewModel;
+package com.traininapp.View;
+
+import com.traininapp.adapter.IStatistic;
 
 import java.util.ArrayList;
 import java.util.Date;
