+++ TrainingApp/app/src/main/java/com/traininapp/adapter/IStatistic.java
@@ -1,10 +1,9 @@
-package com.traininapp.viewModel;
+package com.traininapp.adapter;
 
 public interface IStatistic {
 
         int TYPE_GOALSTAT = 101;
         int TYPE_STATISTIC = 102;
-
 
         int getType();
 }
