+++ TrainingApp/app/src/main/res/layout/activity_create_routine.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".viewModel.CreateRoutine">
+    tools:context=".View.CreateRoutine">
 
     <Button
         android:id="@+id/btnAddExerciseID"
